requirements.txt
SQLAlchemy
python-dotenv
loguru
pydantic
fastapi
uvicorn
passlib
pyjwt
python-multipart
asyncpg
minio